+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## HEAD
+
+* Adding `Document#done?` to check if a document is complete
+
 ## 0.2.0
 
 * Handling HTTP and API errors and exposing the error messages